Move ParameterMixin to extensions
